1) Error sending mails to deleted users (might be an issue, that already exists since a long time).  in order to reproduce you have to have deleted users (field deleted is set to 1 I think in mdl_user)
Executing cron job displays following error:
</pre>Execute adhoc task: mod_booking\task\send_confirmation_mails
... started 11:55:29. Current memory use 15.1MB.
++ Can not send email to deleted user: 84 ++
* line 5731 of /lib/moodlelib.php: call to debugging()
* line 37 of /mod/booking/classes/task/send_confirmation_mails.php: call
to email_to_user()
* line 81 of /lib/cronlib.php: call to
mod_booking\task\send_confirmation_mails->execute()
* line 81 of /admin/cron.php: call to cron_run()
... used 0 dbqueries
... used 0.00063014030456543 seconds
Adhoc task failed: mod_booking\task\send_confirmation_mails,Coding error
detected, it must be fixed by a programmer: Confirmation email was not
sent
Backtrace:
* line 81 of /lib/cronlib.php: call to
mod_booking\task\send_confirmation_mails->execute()
* line 81 of /admin/cron.php: call to cron_run()

Check is missing if user is deleted. Also deleted users should be
cleared from any booking_answer when the delete user event occurs.

+++ classes/task/send_confirmation_mails.php
@@ -32,15 +32,18 @@ class send_confirmation_mails extends \core\task\adhoc_task {
         $taskdata = $this->get_custom_data();
 
         if ($taskdata != null) {
-            if (!email_to_user($taskdata->userto, $taskdata->userfrom, $taskdata->subject,
-                    $taskdata->messagetext, $taskdata->messagehtml, $taskdata->attachment,
-                    $taskdata->attachname)) {
-                throw new \coding_exception('Confirmation email was not sent');
-            } else {
-                $search = str_replace($CFG->tempdir . '/', '', $taskdata->attachment);
-                if ($DB->count_records_select('task_adhoc', "customdata LIKE '%$search%'") == 1) {
-                    if (!empty($taskdata->attachment)) {
-                        unlink($taskdata->attachment);
+            $userdata = $DB->get_record('user', array('id' => $taskdata->userto->id));
+            if (!$userdata->deleted) {
+                if (!email_to_user($taskdata->userto, $taskdata->userfrom, $taskdata->subject,
+                        $taskdata->messagetext, $taskdata->messagehtml, $taskdata->attachment,
+                        $taskdata->attachname)) {
+                    throw new \coding_exception('Confirmation email was not sent');
+                } else {
+                    $search = str_replace($CFG->tempdir . '/', '', $taskdata->attachment);
+                    if ($DB->count_records_select('task_adhoc', "customdata LIKE '%$search%'") == 1) {
+                        if (!empty($taskdata->attachment)) {
+                            unlink($taskdata->attachment);
+                        }
                     }
                 }
             }
